requirements.txt
# Base ----------------------------------------
sqlalchemy==2.0.17
pymysql==1.1.1
pandas==2.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
spotipy==2.24.0

numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.8.0
ipyleaflet==0.17.0
sympy==1.12

# Machine learning ----------------------------
scikit-learn==1.3.1